[Update] New file to print all directories of path, and update the _getenv function to did it dinamically
